update dashboard page and refactor related code

+++ backend/src/main/java/com/nvb/fin_flow/service/impl/CategoryServiceImpl.java
@@ -1,32 +1,32 @@
 package com.nvb.fin_flow.service.impl;
 
 import com.nvb.fin_flow.dto.request.CategoryCreationRequest;
+import com.nvb.fin_flow.dto.response.CategoryPageableResponse;
 import com.nvb.fin_flow.dto.response.CategoryResponse;
 import com.nvb.fin_flow.entity.Category;
 import com.nvb.fin_flow.entity.User;
-import com.nvb.fin_flow.enums.CategoryType;
 import com.nvb.fin_flow.exception.AppException;
 import com.nvb.fin_flow.exception.ErrorCode;
 import com.nvb.fin_flow.mapper.CategoryMapper;
 import com.nvb.fin_flow.repository.CategoryRepository;
-import com.nvb.fin_flow.repository.UserRepository;
 import com.nvb.fin_flow.repository.projection.CategoryWithIcon;
-import com.nvb.fin_flow.repository.projection.CategoryWithIconAndCreatedBy;
+import com.nvb.fin_flow.repository.projection.CategoryWithIconAndUser;
 import com.nvb.fin_flow.service.CategoryService;
+import com.nvb.fin_flow.service.UserService;
+import com.nvb.fin_flow.specification.CategorySpecifications;
+import com.nvb.fin_flow.utilities.PageableUtility;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.security.authentication.AnonymousAuthenticationToken;
-import org.springframework.security.core.Authentication;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@@ -36,47 +36,32 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
     CategoryRepository categoryRepository;
     CategoryMapper categoryMapper;
-    UserRepository userRepository;
+    UserService userService;
+    PageableUtility pageableUtility;
 
     @Override
-    public Set<CategoryResponse> getCategoriesByUsername(String username, CategoryType type, Map<String, String> params) {
-        List<CategoryWithIconAndCreatedBy> categories = categoryRepository
-                .findByCreatedBy_UsernameAndType(username, type, CategoryWithIconAndCreatedBy.class);
+    public CategoryPageableResponse getCategories(Map<String, String> params) {
+        Pageable page = pageableUtility.getPageable(params.get("page"), null);
 
-        if (params.containsKey("q")) {
-            String keyword = params.get("q").toLowerCase();
-            categories = categories.stream()
-                    .filter(c -> c.getName().toLowerCase().contains(keyword))
-                    .toList();
-        }
+        Specification<Category> spec =
+                Specification.allOf(CategorySpecifications.belongToUser(params.get("username")),
+                        CategorySpecifications.nameContains(params.get("q")),
+                        CategorySpecifications.typeContains(params.get("type")));
 
-        if (params.containsKey("type")) {
-            String typeParam = params.get("type").toUpperCase();
-            if (Arrays.stream(CategoryType.values())
-                    .anyMatch(p -> p.name().equals(typeParam))) {
+        Page<CategoryWithIconAndUser> categories = categoryRepository
+                .findBy(spec, q ->q.as(CategoryWithIconAndUser.class).page(page));
 
-                categories = categories.stream()
-                        .filter(c -> c.getType().equalsIgnoreCase(typeParam))
-                        .toList();
-            }
-        }
-
-        return categories.stream()
-                .map(categoryMapper::toResponse)
-                .collect(Collectors.toSet());
+        return CategoryPageableResponse.builder()
+                .totalPages(categories.getTotalPages())
+                .categoryResponses(new LinkedHashSet<>(categories.getContent().stream().map(categoryMapper::toResponse).collect(Collectors.toList())))
+                .build();
     }
 
     @Override
-    public CategoryResponse addOrCategory(CategoryCreationRequest categoryCreationRequest) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication == null || !authentication.isAuthenticated() || authentication instanceof AnonymousAuthenticationToken) {
-            throw new AppException(ErrorCode.UNAUTHENTICATED);
-        }
-        String username = authentication.getName();
-        User createdBy = userRepository.findByUsername(username)
-                .orElseThrow(() -> new AppException(ErrorCode.ENTITY_NOT_EXISTED, Map.of("entity", "User")));
+    public CategoryResponse addOrUpdateCategory(CategoryCreationRequest categoryCreationRequest) {
         Category category = categoryMapper.toEntity(categoryCreationRequest);
-        category.setCreatedBy(createdBy);
+        User user = userService.getCurrentUser();
+        category.setUser(user);
         try{
             categoryRepository.save(category);
         }catch (DataIntegrityViolationException e){
@@ -85,9 +70,17 @@ public class CategoryServiceImpl implements CategoryService {
         return categoryMapper.toResponse(category);
     }
 
+
     @Override
     public void deleteCategory(String id) {
         try{
+            Category category = categoryRepository.findById(id).orElse(null);
+            if(category == null){
+                throw new AppException(ErrorCode.ENTITY_NOT_EXISTED, Map.of("entity", "Category"));
+            }
+            if(!category.getUser().getUsername().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+                throw new AppException(ErrorCode.DO_NOT_HAVE_PERMISSION);
+            }
             categoryRepository.deleteById(id);
         }catch (DataIntegrityViolationException e){
             throw new AppException(ErrorCode.SQL_EXCEPTION, Map.of("ms", e.getCause().getMessage().split("\\[")[0].trim()));
